% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkagesim-package.R
\docType{package}
\name{linkagesim-package}
\alias{linkagesim}
\alias{linkagesim-package}
\title{linkagesim: simulacao e validacao de record linkage para bases do SUS}
\description{
Gera bancos sinteticos com inconsistencias controladas a partir de uma base
de pessoas conhecida, executa o relacionamento probabilistico e mede
sensibilidade, valor preditivo positivo e F1 contra a verdade conhecida.
}
\details{
O fluxo tipico e:
\enumerate{
  \item \code{\link{ln_gerar_base_comum}} e \code{\link{ln_criar_banco}} para
    montar os dois bancos sinteticos;
  \item \code{\link{ln_pipeline}} para executar o linkage;
  \item \code{\link{ln_avaliar}} e \code{\link{ln_calibrar_limiar}} para medir
    o desempenho e escolher o ponto de corte.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USUARIO/linkagesim}
  \item Report bugs at \url{https://github.com/USUARIO/linkagesim/issues}
}

}
\author{
\strong{Maintainer}: PRIMEIRO SOBRENOME \email{seu.email@exemplo.br} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
