% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliacao.R
\name{ln_avaliar}
\alias{ln_avaliar}
\title{Metricas de validacao da simulacao}
\usage{
ln_avaliar(pares, A, B, prefixo_comum = "^COMUM_")
}
\arguments{
\item{pares}{saida de \code{ln_decidir()}.}

\item{A, B}{tabelas de pessoas.}

\item{prefixo_comum}{expressao regular que identifica as pessoas presentes
nos dois bancos.}
}
\value{
\code{data.table} com sensibilidade, valor preditivo positivo e F1.
}
\description{
Metricas de validacao da simulacao
}
