% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{ln_criar_chaves}
\alias{ln_criar_chaves}
\title{Cria as chaves de bloco}
\usage{
ln_criar_chaves(p)
}
\arguments{
\item{p}{\code{data.table} de pessoas, saida de \code{ln_perfil_pessoa()}.}
}
\value{
O mesmo \code{data.table} com as colunas \code{k_*}.
}
\description{
Cria as chaves de bloco
}
