% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulacao.R
\name{ln_formatar_data}
\alias{ln_formatar_data}
\title{Formata datas em layouts variados}
\usage{
ln_formatar_data(datas, p_vazia, formatos)
}
\arguments{
\item{datas}{vetor \code{Date}.}

\item{p_vazia}{proporcao de datas devolvidas em branco.}

\item{formatos}{vetor nomeado com os formatos e suas proporcoes.}
}
\value{
Vetor de caracteres.
}
\description{
Formata datas em layouts variados
}
