% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{ln_gerar_candidatos}
\alias{ln_gerar_candidatos}
\title{Gera os pares candidatos pela uniao dos passos de blocking}
\usage{
ln_gerar_candidatos(
  A,
  B,
  passos = ln_passos_padrao(),
  max_pares_bloco = 5e+05,
  verbose = TRUE
)
}
\arguments{
\item{A, B}{tabelas de pessoas com as chaves criadas por \code{ln_criar_chaves()}.}

\item{passos}{lista de passos, de \code{ln_passos_padrao()}.}

\item{max_pares_bloco}{teto de pares por chave; blocos maiores (nomes muito
frequentes) sao ignorados naquele passo e avisados.}

\item{verbose}{imprimir o progresso.}
}
\value{
\code{data.table} com as colunas \code{ida} e \code{idb}.
}
\description{
Gera os pares candidatos pela uniao dos passos de blocking
}
