% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulacao.R
\name{ln_gerar_cns_valido}
\alias{ln_gerar_cns_valido}
\title{Gera CNS provisorios validos pelo modulo 11}
\usage{
ln_gerar_cns_valido(n)
}
\arguments{
\item{n}{quantidade de CNS a gerar.}
}
\value{
Vetor de caracteres com 15 digitos.
}
\description{
Gera CNS provisorios validos pelo modulo 11
}
