% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escore.R
\name{ln_par_escore}
\alias{ln_par_escore}
\title{Parametros do escore Fellegi-Sunter}
\usage{
ln_par_escore(
  m_nome = 0.92,
  u_nome = 0.004,
  m_mae = 0.88,
  u_mae = 0.008,
  m_data = 0.93,
  u_data = 1/3650,
  m_cns = 0.85,
  u_cns = 1e-06,
  corte_nome = 0.85,
  corte_mae = 0.85,
  corte_data = 0.8,
  corte_cns = 0.6
)
}
\arguments{
\item{m_nome, u_nome, m_mae, u_mae, m_data, u_data, m_cns, u_cns}{probabilidades m e
u de cada campo.}

\item{corte_nome, corte_mae, corte_data, corte_cns}{similaridade a partir da
qual a concordancia e considerada total; abaixo dela a contribuicao e
interpolada linearmente ate o peso de discordancia.}
}
\value{
Lista de parametros.
}
\description{
Parametros do escore Fellegi-Sunter
}
