% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulacao.R
\name{ln_par_ruido}
\alias{ln_par_ruido}
\title{Parametros de ruido da simulacao}
\usage{
ln_par_ruido(
  p_nome_vazio = 0.03,
  p_nome_abrev = 0.07,
  p_nome_typo = 0.05,
  p_mae_vazio = 0.08,
  p_mae_abrev = 0.17,
  p_mae_typo = 0.05,
  p_data_vazia = 0.05,
  p_cns_trocado = 0.08,
  p_cns_vazio = 0.05,
  p_cns_invalido = 0.03,
  p_duplicata = 0.05,
  formatos_data = c(`\%d/\%m/\%Y` = 0.3, `\%Y-\%m-\%d` = 0.25, `\%d-\%m-\%Y` = 0.15,
    `\%d\%m\%Y` = 0.1, `\%Y/\%m/\%d` = 0.2)
)
}
\arguments{
\item{p_nome_vazio, p_nome_abrev, p_nome_typo}{taxas para o nome do paciente.}

\item{p_mae_vazio, p_mae_abrev, p_mae_typo}{taxas para o nome da mae.}

\item{p_data_vazia}{proporcao de datas de nascimento em branco.}

\item{p_cns_trocado, p_cns_vazio, p_cns_invalido}{taxas para o CNS.}

\item{p_duplicata}{proporcao de registros duplicados dentro do banco.}

\item{formatos_data}{vetor nomeado com os formatos de data e suas proporcoes.}
}
\value{
Lista de parametros.
}
\description{
Reune num unico objeto todas as taxas de inconsistencia injetadas nos bancos
sinteticos, para que possam ser variadas de forma explicita.
}
\examples{
par <- ln_par_ruido(p_mae_vazio = 0.20)
}
