% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escore.R
\name{ln_pesos}
\alias{ln_pesos}
\title{Pesos de concordancia e discordancia de Fellegi-Sunter}
\usage{
ln_pesos(m, u)
}
\arguments{
\item{m}{probabilidade de concordancia dado que o par e verdadeiro.}

\item{u}{probabilidade de concordancia por acaso.}
}
\value{
Vetor nomeado com \code{mais} (log2(m/u)) e \code{menos}
  (log2((1-m)/(1-u))).
}
\description{
Pesos de concordancia e discordancia de Fellegi-Sunter
}
